pizza_project/pizza/requirements.txt
Babel==2.4.0
Django==1.11.4
Pillow==4.2.0
Pygments==2.2.0
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.19
certifi==2017.4.17
chardet==3.0.4
decorator==4.1.2
django-admin-sortable2==0.6.13
django-celery==3.2.1
django-extensions==1.7.9
django-filter==1.0.4
django-jet==1.0.6
django-model-utils==3.0.0
django-parler==1.8
django-phonenumber-field==1.3.0
django-redis==4.8.0
django-rosetta==0.7.13
djangorestframework==3.6.3
easy-thumbnails==2.4.1
future==0.16.0
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
kombu==3.0.37
microsofttranslator==0.8
olefile==0.44
pexpect==4.2.1
phonenumberslite==8.5.2
pickleshare==0.7.4
polib==1.0.8
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.2
python-telegram-bot==8.0
pytz==2017.2
redis==2.10.5
requests==2.18.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
typing==3.6.1
urllib3==1.21.1
vine==1.1.4
wcwidth==0.1.7
